+++ stdlib/importlib/resources/__init__.pyi
@@ -38,23 +38,7 @@ if sys.version_info >= (3, 13):
 
     __all__ += ["Anchor"]
 
-    def open_binary(anchor: Anchor, *path_names: StrPath) -> BinaryIO: ...
-    @overload
-    def open_text(
-        anchor: Anchor, *path_names: Unpack[tuple[StrPath]], encoding: str | None = "utf-8", errors: str | None = "strict"
-    ) -> TextIOWrapper: ...
-    @overload
-    def open_text(anchor: Anchor, *path_names: StrPath, encoding: str | None, errors: str | None = "strict") -> TextIOWrapper: ...
-    def read_binary(anchor: Anchor, *path_names: StrPath) -> bytes: ...
-    @overload
-    def read_text(
-        anchor: Anchor, *path_names: Unpack[tuple[StrPath]], encoding: str | None = "utf-8", errors: str | None = "strict"
-    ) -> str: ...
-    @overload
-    def read_text(anchor: Anchor, *path_names: StrPath, encoding: str | None, errors: str | None = "strict") -> str: ...
-    def path(anchor: Anchor, *path_names: StrPath) -> AbstractContextManager[Path]: ...
-    def is_resource(anchor: Anchor, *path_names: StrPath) -> bool: ...
-    def contents(anchor: Anchor, *path_names: StrPath) -> Iterator[str]: ...
+    from importlib.resources._functional import contents, is_resource, open_binary, open_text, path, read_binary, read_text
 
 else:
     def open_binary(package: Package, resource: Resource) -> BinaryIO: ...
